Break ImportCsvButton into functions
