+++ src/components/Admin/ImportCsvButton/index.jsx
@@ -3,7 +3,7 @@ import { useNotify } from 'react-admin';
 import { ImportButton } from 'react-admin-import-csv';
 import PropTypes from 'prop-types';
 
-import { adminOfficeRoles } from 'constants/userRoles';
+import { checkRequiredFields, checkTelephone, parseRoles } from './validation';
 
 // Note: There is not a test file or story for ImportCsvButton beacuse this component HAS to render within a react-admin app
 const ImportCsvButton = (props) => {
@@ -11,27 +11,17 @@ const ImportCsvButton = (props) => {
   const { resource } = props;
 
   const validateRow = async (row) => {
-    // Verify we have all required fields
-    if (!(row.transportationOfficeId && row.firstName && row.lastName && row.roles && row.email && row.telephone)) {
-      const err = new Error(
-        `Validation Error: Row does not contain all required fields.
-        Required fields are firstName, lastName, email, telephone, roles, and transportationOfficeId \n
-        Row Information: ${Object.values(row)}`,
-      );
-      notify(err.message);
-      throw err;
-    }
-
-    // Verify the phone format
-    const regex = /^[2-9]\d{2}-\d{3}-\d{4}$/;
-    if (!regex.test(row.telephone)) {
-      const err = new Error(
-        `Validation Error: Row contains improperly formatted telephone number. Required format is xxx-xxx-xxxx. \n
-        Row Information: ${Object.values(row)}`,
-      );
-      notify(err.message);
-      throw err;
-    }
+    // Verify we have all required fields and that the telephone is valid
+    const validation = [checkRequiredFields, checkTelephone];
+    validation.forEach((check) => {
+      try {
+        // eslint-disable-next-line react/jsx-props-no-spreading
+        check({ ...row });
+      } catch (err) {
+        notify(`${err.message} \n Row Information: ${Object.values(row)}`);
+        throw err;
+      }
+    });
 
     return row;
   };
@@ -40,39 +30,17 @@ const ImportCsvButton = (props) => {
     const alteredRows = [];
     rows.forEach((row) => {
       const copyOfRow = row;
-      if (row.roles) {
-        const rolesArray = [];
-
-        // Parse roles from string at ","
-        const parsedRoles = row.roles.split(',');
-        parsedRoles.forEach((parsedRole) => {
-          let roleNotFound = true;
-          // Remove any whitespace in the role string
-          const role = parsedRole.replaceAll(/\s/g, '');
-          adminOfficeRoles.forEach((adminOfficeRole) => {
-            if (adminOfficeRole.roleType === role) {
-              rolesArray.push(adminOfficeRole);
-              roleNotFound = false;
-            }
-          });
-          if (roleNotFound) {
-            const err = new Error(
-              `Processing Error: Invalid roles provided for row. \n Row Information: ${Object.values(row)}`,
-            );
-            notify(err.message);
-            throw err;
-          }
-        });
-        copyOfRow.roles = rolesArray;
-      } else {
-        const err = new Error(
-          `Processing Error: Unable to parse roles for row. \n Row Information: ${Object.values(row)}`,
-        );
-        notify(err.message);
+      try {
+        const parsedRolesArray = parseRoles(row.roles);
+        copyOfRow.roles = parsedRolesArray;
+      } catch (err) {
+        notify(`${err.message} \n Row Information: ${Object.values(row)}`);
         throw err;
       }
+
       alteredRows.push(copyOfRow);
     });
+
     return alteredRows;
   };
